server/requirements.txt
blinker==1.9.0
click==8.2.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
fsspec==2025.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
pillow==11.2.1
setuptools==80.7.1
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
typing_extensions==4.13.2
Werkzeug==3.1.3
